+++ src/services/firebase/maintenanceService.js
@@ -1,4 +1,4 @@
-import { app } from '../../firebase';
+import { app } from './config';
 import { USE_MOCK } from './mockDb';
 
 const LOCAL_STORAGE_EQUIPMENTS_KEY = 'nexa_maintenance_equipments';